Fix secret command layer ownership
