+++ docs/webapp-details.md
@@ -5,7 +5,7 @@ Overflow from `packages/webapp/CLAUDE.md`. Each section is the deep reference fo
 ## Kernel Host
 
 - Path: `packages/webapp/src/kernel/`
-- `host.ts` — `createKernelHost(config)` factory: single boot sequence for every float (orchestrator, lick-manager, agent-bridge, tray subs, cone bootstrap, BshWatchdog, `/proc` mount). `node-rest` opens the `/licks-ws` bridge; extension-delegate leaders route licks through the tray worker. Float discriminator: `resolveFloatTopology()` in `core/float-topology.ts`. `kernel-worker.ts` is the DedicatedWorker entry.
+- `host.ts` — `createKernelHost(config)` factory: single boot sequence for every float (orchestrator, lick-manager, agent-bridge, tray subs, cone bootstrap, BshWatchdog, `/proc` mount). `node-rest` opens the `/licks-ws` bridge; extension-delegate leaders route licks through the tray worker. Float discriminator: `resolveFloatTopology()` in `shell/float-topology.ts` (`core/float-topology.ts` is a compatibility re-export). `kernel-worker.ts` is the DedicatedWorker entry.
 - `process-manager.ts` tracks every long-running async unit; pids uint32 from 1024+; `signal(pid, sig)` honors SIGINT/SIGTERM/SIGKILL/SIGSTOP/SIGCONT. `proc-mount.ts` serves read-only `/proc` (`cat /proc/<pid>/{status,cmdline,cwd,stat}`).
 - `realm/` hard-killable runner. `runInRealm()` spawns per-task `DedicatedWorker`; SIGKILL → exit 137. JS runs in kernel worker via `createJsWorkerRealm()`; `realm-host` proxies `vfs`/`exec`/`fetch`. Node shims in `realm/helpers/` (barrel: `js-realm-helpers.ts`). Sync-XHR bridge (`readFileSync`/`writeFileSync`, `child_process.execSync`/`execFileSync`/`spawnSync`) via `realm/sync-{xhr,fs-*,exec-*}.ts` + `ui/sync-fs-sw-handler.ts` with per-realm capability tokens.
 - Deep reference: `docs/kernel/process-model.md`.
